+++ packages/xchain-bitcoin/__tests__/client.test.ts
@@ -50,6 +50,14 @@ describe('BitcoinClient Test', () => {
     expect(result).toEqual(addyOnePath0)
   })
 
+  it('should not throw on a client without a phrase', () => {
+    expect(() => {
+      new Client({
+        network: Network.Testnet,
+      })
+    }).not.toThrow()
+  })
+
   it('should throw an error for setting a bad phrase', () => {
     expect(() => btcClient.setPhrase('cat')).toThrow()
   })
@@ -373,5 +381,16 @@ describe('BitcoinClient Test', () => {
     btcClient.setPhrase(phraseTwo)
     expect(btcClient.getAddress(0)).toEqual(addyThreePath0)
     expect(btcClient.getAddress(1)).toEqual(addyThreePath1)
+  })
+
+  it('should broadcast a deposit to thorchain inbound address', async () => {
+    btcClient.setNetwork(Network.Testnet)
+    btcClient.setPhrase(phraseOne)
+    const txHash = await btcClient.deposit({
+      asset: AssetBTC,
+      amount: baseAmount(2223),
+      memo: '=:THOR.RUNE:tthor1puhn8fclwvmmzh7uj7546wnxz5h3zar8e66sc5',
+    })
+    expect(txHash).toEqual('mock-txid')
   })
 })