+++ packages/xchain-ethereum/__mocks__/thornode-api.ts
@@ -10,6 +10,12 @@ type InboundAddress = {
   router?: string
 }
 
+type MimirDetails = {
+  HALTTRADING: number
+  HALTETHTRADING: number
+  HALTETHCHAIN: number
+}
+
 export const mock_thornode_inbound_addresses_success = (url: string, result: InboundAddress[]) => {
   nock(url)
     .get(`/thorchain/inbound_addresses`)
@@ -24,4 +30,11 @@ export const mock_thornode_inbound_addresses_fail = (url: string) => {
     .reply(500, {
       error: 'somthing bad happened',
     })
+}
+
+export const mock_thornode_mimir = (url: string, result: MimirDetails) => {
+  nock(url)
+    .get('/thorchain/mimir')
+    .query((_) => true)
+    .reply(200, result)
 }