+++ packages/xchain-terra/__e2e__/client.e2e.ts
@@ -1,17 +1,33 @@
 import { Network, Tx } from '@xchainjs/xchain-client'
-import { Chain, baseAmount } from '@xchainjs/xchain-util'
+import { Chain, assetAmount, assetToBase, baseAmount, delay } from '@xchainjs/xchain-util'
 
-import { Client } from '../src'
+/**
+ * e2e test
+ *
+ * How to run?
+ *
+ * (1) Add `phrase` to `.env` file (see `.env.sample`)
+ * (2) Run `yarn e2e` from `src/packages/xhain-client`
+ *
+ * Note: In case it fails to run all e2e at once, run it one by one by adding `only` to a test (@see https://jestjs.io/docs/setup-teardown#general-advice)
+ *
+ */
+
+require('dotenv').config()
+
+import { AssetLUNA, AssetUST, Client, TERRA_DECIMAL, getEstimatedFee } from '../src'
+
+const phrase = process.env.PHRASE
 
 describe('Client Test', () => {
   let terraClient: Client
-  const phrase = process.env.PHRASE
 
   beforeEach(() => {
     terraClient = new Client({ phrase, network: Network.Testnet })
   })
 
-  afterEach(() => {
+  afterEach(async () => {
+    await delay(200)
     terraClient.purgeClient()
   })
 
@@ -34,14 +50,14 @@ describe('Client Test', () => {
       fail()
     }
   })
-  it('should get KRT transaction details', async () => {
+  it('should get KRW transaction details', async () => {
     try {
       terraClient.setNetwork(Network.Testnet)
       const txHash = 'A8057659C5F189B91E3D85794479D80A78B5A4F6D4C5E6CAC01A3FADB274332F'
       const txDetails = await terraClient.getTransactionData(txHash)
       console.log(JSON.stringify(txDetails))
-      expect(txDetails.from[0].asset?.symbol).toEqual('KRT')
-      expect(txDetails.from[0].asset?.ticker).toEqual('KRT')
+      expect(txDetails.from[0].asset?.symbol).toEqual('KRW')
+      expect(txDetails.from[0].asset?.ticker).toEqual('KRW')
       expect(txDetails.from[0].from).toEqual('terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv')
       expect(txDetails.from[0].amount.amount().toFixed()).toEqual('1000000')
       expect(txDetails.to[0].to).toEqual('terra1ltnkx0mv7lf2rca9f8w740ashu93ujughy4s7p')
@@ -55,7 +71,7 @@ describe('Client Test', () => {
   })
   it('should get address balances', async () => {
     try {
-      terraClient.setNetwork('testnet' as Network)
+      terraClient.setNetwork(Network.Testnet)
 
       const address = 'terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv'
       const balances = await terraClient.getBalance(address)
@@ -79,11 +95,37 @@ describe('Client Test', () => {
       fail()
     }
   })
-  it('should get fees', async () => {
+  it('get fees - testnet', async () => {
     try {
-      terraClient.setNetwork(Network.Mainnet)
+      terraClient.setNetwork(Network.Testnet)
 
-      const fees = await terraClient.getFees()
+      const fees = await terraClient.getFees({
+        sender: 'terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv',
+        recipient: 'terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv',
+        feeAsset: AssetLUNA,
+        asset: AssetUST,
+        amount: assetToBase(assetAmount(0.001)),
+      })
+
+      console.log(JSON.stringify(fees.fast.amount()))
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+  })
+
+  it('get fees w/ memo - testnet', async () => {
+    try {
+      terraClient.setNetwork(Network.Testnet)
+
+      const fees = await terraClient.getFees({
+        sender: 'terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv',
+        recipient: 'terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv',
+        feeAsset: AssetUST,
+        asset: AssetUST,
+        amount: assetToBase(assetAmount(0.01)),
+        memo: '+:TERRA.UST',
+      })
 
       console.log(JSON.stringify(fees))
     } catch (error) {
@@ -107,36 +149,36 @@ describe('Client Test', () => {
 
       //Inputs
       expect(tx.from[0].from).toEqual('terra1fex9f78reuwhfsnc8sun6mz8rl9zwqh03fhwf3')
-      expect(tx.from[0].asset?.symbol).toEqual('KRT')
+      expect(tx.from[0].asset?.symbol).toEqual('KRW')
       expect(tx.from[0].amount.amount().toFixed()).toEqual('1')
       expect(tx.from[1].from).toEqual('terra1l6834ha5h5l5dxkr0vl82hjcucfht3gpwlflq6')
-      expect(tx.from[1].asset?.symbol).toEqual('KRT')
+      expect(tx.from[1].asset?.symbol).toEqual('KRW')
       expect(tx.from[1].amount.amount().toFixed()).toEqual('41948000000')
       expect(tx.from[2].from).toEqual('terra1fex9f78reuwhfsnc8sun6mz8rl9zwqh03fhwf3')
-      expect(tx.from[2].asset?.symbol).toEqual('KRT')
+      expect(tx.from[2].asset?.symbol).toEqual('KRW')
       expect(tx.from[2].amount.amount().toFixed()).toEqual('3000000000')
       expect(tx.from[3].from).toEqual('terra1dtcvy52ma8gl5f2lx4klpqzcyfw90tujp66yp7')
-      expect(tx.from[3].asset?.symbol).toEqual('KRT')
+      expect(tx.from[3].asset?.symbol).toEqual('KRW')
       expect(tx.from[3].amount.amount().toFixed()).toEqual('17000000000')
       expect(tx.from[4].from).toEqual('terra1fex9f78reuwhfsnc8sun6mz8rl9zwqh03fhwf3')
-      expect(tx.from[4].asset?.symbol).toEqual('KRT')
+      expect(tx.from[4].asset?.symbol).toEqual('KRW')
       expect(tx.from[4].amount.amount().toFixed()).toEqual('25000000000')
 
       //outputs
       expect(tx.to[0].to).toEqual('terra1fex9f78reuwhfsnc8sun6mz8rl9zwqh03fhwf3')
-      expect(tx.to[0].asset?.symbol).toEqual('KRT')
+      expect(tx.to[0].asset?.symbol).toEqual('KRW')
       expect(tx.to[0].amount.amount().toFixed()).toEqual('1')
       expect(tx.to[1].to).toEqual('terra1fd3yy3dlg6gm0fkapyqzcunl26zkwdnal0y2cq')
-      expect(tx.to[1].asset?.symbol).toEqual('KRT')
+      expect(tx.to[1].asset?.symbol).toEqual('KRW')
       expect(tx.to[1].amount.amount().toFixed()).toEqual('41948000000')
       expect(tx.to[2].to).toEqual('terra1l6834ha5h5l5dxkr0vl82hjcucfht3gpwlflq6')
-      expect(tx.to[2].asset?.symbol).toEqual('KRT')
+      expect(tx.to[2].asset?.symbol).toEqual('KRW')
       expect(tx.to[2].amount.amount().toFixed()).toEqual('3000000000')
       expect(tx.to[3].to).toEqual('terra1fd3yy3dlg6gm0fkapyqzcunl26zkwdnal0y2cq')
-      expect(tx.to[3].asset?.symbol).toEqual('KRT')
+      expect(tx.to[3].asset?.symbol).toEqual('KRW')
       expect(tx.to[3].amount.amount().toFixed()).toEqual('17000000000')
       expect(tx.to[4].to).toEqual('terra1nqnmgh6vkrcq8e4sscjxwnyu740g5xuxlk9edr')
-      expect(tx.to[4].asset?.symbol).toEqual('KRT')
+      expect(tx.to[4].asset?.symbol).toEqual('KRW')
       expect(tx.to[4].amount.amount().toFixed()).toEqual('25000000000')
     } catch (error) {
       console.error(error)
@@ -158,45 +200,277 @@ describe('Client Test', () => {
       expect(tx.to.length).toEqual(2)
 
       expect(tx.from[0].from).toEqual('terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv')
-      expect(tx.from[0].asset?.symbol).toEqual('KRT')
+      expect(tx.from[0].asset?.symbol).toEqual('KRW')
       expect(tx.from[0].amount.amount().toFixed()).toEqual('1000000')
       expect(tx.to[0].to).toEqual('terra1ltnkx0mv7lf2rca9f8w740ashu93ujughy4s7p')
-      expect(tx.to[0].asset?.symbol).toEqual('KRT')
+      expect(tx.to[0].asset?.symbol).toEqual('KRW')
       expect(tx.to[0].amount.amount().toFixed()).toEqual('1000000')
 
       expect(tx.from[1].from).toEqual('terra1saynp5x60tr03sy4awr2rzt3wgmqrqahuahccv')
-      expect(tx.from[1].asset?.symbol).toEqual('KRT')
+      expect(tx.from[1].asset?.symbol).toEqual('KRW')
       expect(tx.from[1].amount.amount().toFixed()).toEqual('1000000')
       expect(tx.to[1].to).toEqual('terra1j6fey5tl70k9fvrv7mea7ahfr8u2yv7l23w5e6')
-      expect(tx.to[1].asset?.symbol).toEqual('KRT')
+      expect(tx.to[1].asset?.symbol).toEqual('KRW')
       expect(tx.to[1].amount.amount().toFixed()).toEqual('1000000')
     } catch (error) {
       console.error(error)
       fail()
     }
   })
-  it('should transfer UST with a memo', async () => {
+
+  it('transfers LUNA, w/ memo, fee LUNA by default', async () => {
+    console.log('### transfer LUNA (fee: LUNA)')
     try {
       terraClient.setNetwork(Network.Testnet)
       console.log(terraClient.getAddress(0))
       console.log(terraClient.getAddress(1))
       const recipient = terraClient.getAddress(1)
-      const hashcode = await terraClient.transfer({
+      const txHash = await terraClient.transfer({
         walletIndex: 0,
-        asset: {
-          chain: Chain.Terra,
-          symbol: 'UST',
-          ticker: 'UST',
-          synth: false,
-        },
-        amount: baseAmount('10000'), // 0.01 UST
+        asset: AssetLUNA,
+        amount: baseAmount(1, TERRA_DECIMAL),
         recipient,
-        memo: 'memo test',
+        memo: 'fee-luna',
       })
-      expect(hashcode).toBeDefined()
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
     } catch (error) {
       console.error(error)
       fail()
     }
+    console.log('### END transfer LUNA (fee: LUNA)')
+  })
+
+  it('transfers LUNA, w/ memo, fee UST', async () => {
+    console.log('### transfer LUNA (fee: UST)')
+    try {
+      terraClient.setNetwork(Network.Testnet)
+      console.log(terraClient.getAddress(0))
+      console.log(terraClient.getAddress(1))
+      const recipient = terraClient.getAddress(1)
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset: AssetLUNA,
+        amount: baseAmount(1, TERRA_DECIMAL),
+        recipient,
+        memo: 'fee-luna',
+        feeAsset: AssetUST,
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer LUNA (fee: UST)')
+  })
+
+  it('transfers LUNA, w/ memo, estimated fee UST ', async () => {
+    console.log('### transfer LUNA (estimated fee: UST)')
+    try {
+      const network = Network.Testnet
+      terraClient.setNetwork(network)
+
+      const asset = AssetLUNA
+      const feeAsset = AssetUST
+      const recipient = terraClient.getAddress(1)
+      const memo = 'est-fee-ust'
+      const amount = baseAmount(1, TERRA_DECIMAL)
+      const { chainID, cosmosAPIURL } = terraClient.getConfig()
+      const { amount: feeAmount, gasLimit } = await getEstimatedFee({
+        chainId: chainID,
+        cosmosAPIURL,
+        sender: terraClient.getAddress(0),
+        recipient,
+        amount,
+        asset,
+        feeAsset,
+        network,
+      })
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset,
+        amount,
+        recipient,
+        memo,
+        feeAsset,
+        feeAmount,
+        gasLimit,
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer LUNA (estimated fee: UST)')
+  })
+
+  it('transfers UST, w/ memo , fee UST by default', async () => {
+    console.log('### transfer UST (fee: UST)')
+    try {
+      terraClient.setNetwork(Network.Testnet)
+      const recipient = terraClient.getAddress(1)
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset: AssetUST,
+        amount: assetToBase(assetAmount(0.01, TERRA_DECIMAL)), // 0.01 UST
+        recipient,
+        memo: 'fee-ust',
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer UST (fee: UST)')
+  })
+
+  it('transfers UST, w/ memo, fee LUNA ', async () => {
+    console.log('### transfer UST (fee: LUNA)')
+    try {
+      terraClient.setNetwork(Network.Testnet)
+      const recipient = terraClient.getAddress(1)
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset: AssetUST,
+        amount: assetToBase(assetAmount(0.01, TERRA_DECIMAL)), // 0.01 UST
+        recipient,
+        memo: 'fee-luna',
+        feeAsset: AssetLUNA,
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer UST (fee: LUNA)')
+  })
+
+  it('transfers UST. w/ memo, estimated fee LUNA ', async () => {
+    console.log('### transfer UST (estimated fee: LUNA)')
+    try {
+      const network = Network.Testnet
+      terraClient.setNetwork(network)
+
+      const asset = AssetUST
+      const feeAsset = AssetUST
+      const recipient = terraClient.getAddress(1)
+      const memo = 'est-fee-luna'
+      const amount = assetToBase(assetAmount(0.01, TERRA_DECIMAL))
+      const { chainID, cosmosAPIURL } = terraClient.getConfig()
+      const { amount: feeAmount, gasLimit } = await getEstimatedFee({
+        chainId: chainID,
+        cosmosAPIURL,
+        sender: terraClient.getAddress(0),
+        recipient,
+        amount,
+        asset,
+        feeAsset,
+        network,
+      })
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset,
+        amount,
+        recipient,
+        memo,
+        feeAsset,
+        feeAmount,
+        gasLimit,
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer UST (estimated fee: LUNA)')
+  })
+
+  it('transfers EUT, fee EUT by default ', async () => {
+    console.log('### transfer EUT (fee: EUT)')
+    try {
+      terraClient.setNetwork(Network.Testnet)
+      const recipient = terraClient.getAddress(1)
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset: { chain: Chain.Terra, symbol: 'EUT', ticker: 'EUT', synth: false },
+        amount: assetToBase(assetAmount(0.01, TERRA_DECIMAL)), // 0.01 EUT
+        recipient,
+        memo: 'fee-luna',
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer EUT (fee: EUT)')
+  })
+
+  it('transfers EUT, fee LUNA ', async () => {
+    console.log('### transfer EUT (fee: LUNA)')
+    try {
+      terraClient.setNetwork(Network.Testnet)
+      const recipient = terraClient.getAddress(1)
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset: { chain: Chain.Terra, symbol: 'EUT', ticker: 'EUT', synth: false },
+        amount: assetToBase(assetAmount(0.01, TERRA_DECIMAL)), // 0.01 EUT
+        recipient,
+        memo: 'fee-luna',
+        feeAsset: AssetLUNA,
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer EUT (fee: LUNA)')
+  })
+
+  it('transfers EUT, estimated fee LUNA ', async () => {
+    console.log('### transfer EUT (estimated fee: LUNA)')
+    try {
+      const network = Network.Testnet
+      terraClient.setNetwork(network)
+
+      const asset = { chain: Chain.Terra, symbol: 'EUT', ticker: 'EUT', synth: false }
+      const feeAsset = AssetLUNA
+      const recipient = terraClient.getAddress(1)
+      const memo = 'est-fee-luna'
+      const amount = assetToBase(assetAmount(0.01, TERRA_DECIMAL))
+      const { chainID, cosmosAPIURL } = terraClient.getConfig()
+      const { amount: feeAmount, gasLimit } = await getEstimatedFee({
+        chainId: chainID,
+        cosmosAPIURL,
+        sender: terraClient.getAddress(0),
+        recipient,
+        amount,
+        asset,
+        feeAsset,
+        network,
+      })
+      const txHash = await terraClient.transfer({
+        walletIndex: 0,
+        asset,
+        amount,
+        recipient,
+        memo,
+        feeAsset,
+        feeAmount,
+        gasLimit,
+      })
+      console.info('tx', terraClient.getExplorerTxUrl(txHash))
+      expect(txHash).toBeDefined()
+    } catch (error) {
+      console.error(error)
+      fail()
+    }
+    console.log('### END transfer EUT (estimated fee: EUT)')
   })
 })