+++ packages/xchain-terra/src/client.ts
@@ -1,115 +1,129 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { AccAddress, Coin, Coins, LCDClient, MnemonicKey, MsgMultiSend, MsgSend, TxInfo } from '@terra-money/terra.js'
 import {
+  AccAddress,
+  Coin,
+  Coins,
+  CreateTxOptions,
+  Fee,
+  LCDClient,
+  MnemonicKey,
+  MsgMultiSend,
+  MsgSend,
+  TxInfo,
+} from '@terra-money/terra.js'
+import { BaseXChainClient, DepositParams, FeeType, Network, TxHash, TxType, singleFee } from '@xchainjs/xchain-client'
+import type {
   Balance,
-  BaseXChainClient,
-  FeeType,
   Fees,
-  Network,
   Tx,
   TxFrom,
   TxHistoryParams,
   TxParams,
   TxTo,
-  TxType,
   TxsPage,
   XChainClient,
   XChainClientParams,
 } from '@xchainjs/xchain-client'
-import { Asset, AssetLUNA, Chain, assetToString, baseAmount } from '@xchainjs/xchain-util'
+import { Asset, BaseAmount, Chain, assetToString, baseAmount, getInboundDetails } from '@xchainjs/xchain-util'
 import axios from 'axios'
+import BigNumber from 'bignumber.js'
 
-import { TERRA_DECIMAL } from './const'
-import { getTerraMicroDenom } from './util'
+import { AssetLUNA, TERRA_DECIMAL } from './const'
+import { EstimatedFee } from './types'
+import type { ClientConfig, ClientParams, FeeParams } from './types/client'
+import {
+  coinsToBalances,
+  getDefaultClientConfig,
+  getDefaultRootDerivationPaths,
+  getEstimatedFee,
+  getTerraNativeAsset,
+  getTerraNativeDenom,
+} from './util'
 
-const DEFAULT_CONFIG: Record<Network, TerraClientConfig> = {
-  [Network.Mainnet]: {
-    explorerURL: 'https://finder.terra.money/mainnet',
-    explorerAddressURL: 'https://finder.terra.money/mainnet/address/',
-    explorerTxURL: 'https://finder.terra.money/mainnet/tx/',
-    cosmosAPIURL: 'https://fcd.terra.dev',
-    ChainID: 'columbus-5',
-  },
-  [Network.Stagenet]: {
-    explorerURL: 'https://finder.terra.money/mainnet',
-    explorerAddressURL: 'https://finder.terra.money/mainnet/address/',
-    explorerTxURL: 'https://finder.terra.money/mainnet/tx/',
-    cosmosAPIURL: 'https://fcd.terra.dev',
-    ChainID: 'columbus-5',
-  },
-  [Network.Testnet]: {
-    explorerURL: 'https://finder.terra.money/testnet',
-    explorerAddressURL: 'https://finder.terra.money/testnet/address/',
-    explorerTxURL: 'https://finder.terra.money/testnet/tx/',
-    cosmosAPIURL: 'https://bombay-fcd.terra.dev',
-    ChainID: 'bombay-12',
-  },
-}
-
-export type SearchTxParams = {
-  messageAction?: string
-  messageSender?: string
-  transferSender?: string
-  transferRecipient?: string
-  page?: number
-  limit?: number
-  txMinHeight?: number
-  txMaxHeight?: number
-}
-export type TerraClientConfig = {
-  explorerURL: string
-  explorerAddressURL: string
-  explorerTxURL: string
-  cosmosAPIURL: string
-  ChainID: string
-}
-export type TerraClientParams = {
-  explorerURL?: string
-  explorerAddressURL?: string
-  explorerTxURL?: string
-  cosmosAPIURL?: string
-  ChainID?: string
+/**
+ * Interface for Terra client
+ */
+export interface TerraClient {
+  // `getFees` of `BaseXChainClient` needs to be overridden
+  getFees(params: FeeParams): Promise<Fees>
+  getEstimatedFee(params: FeeParams): Promise<EstimatedFee>
 }
 
 /**
  * Terra Client
  */
-class Client extends BaseXChainClient implements XChainClient {
+class Client extends BaseXChainClient implements XChainClient, TerraClient {
   private lcdClient: LCDClient
-  private config: Record<Network, TerraClientConfig>
+  private config: Record<Network, ClientConfig>
   constructor({
     network = Network.Testnet,
     phrase,
-    rootDerivationPaths = {
-      [Network.Mainnet]: "44'/330'/0'/0/",
-      [Network.Stagenet]: "44'/330'/0'/0/",
-      [Network.Testnet]: "44'/330'/0'/0/",
-    },
+    rootDerivationPaths = getDefaultRootDerivationPaths(),
     explorerURL,
     explorerAddressURL,
     explorerTxURL,
     cosmosAPIURL,
-    ChainID,
-  }: XChainClientParams & TerraClientParams) {
+    chainID,
+  }: XChainClientParams & ClientParams) {
     super(Chain.Terra, { network, rootDerivationPaths, phrase })
-    this.config = { ...DEFAULT_CONFIG, ...{ explorerURL, explorerAddressURL, explorerTxURL, cosmosAPIURL, ChainID } }
+    const defaultClientConfigs = getDefaultClientConfig()
+    const defaultClientConfig = defaultClientConfigs[network]
+    this.config = {
+      ...defaultClientConfigs,
+      // override default config of given network with given values
+      [network]: {
+        ...defaultClientConfig,
+        explorerURL: explorerURL || defaultClientConfig.explorerURL,
+        explorerAddressURL: explorerAddressURL || defaultClientConfig.explorerAddressURL,
+        explorerTxURL: explorerTxURL || defaultClientConfig.explorerTxURL,
+        cosmosAPIURL: cosmosAPIURL || defaultClientConfig.cosmosAPIURL,
+        chainID: chainID || defaultClientConfig.chainID,
+      },
+    }
 
     this.lcdClient = new LCDClient({
-      URL: this.config[this.network].cosmosAPIURL,
-      chainID: this.config[this.network].ChainID,
+      URL: defaultClientConfig.cosmosAPIURL,
+      chainID: defaultClientConfig.chainID,
     })
   }
 
-  async getFees(): Promise<Fees> {
-    const feesArray = (await axios.get(`${this.config[this.network].cosmosAPIURL}/v1/txs/gas_prices`)).data
-    const baseFeeInLuna = baseAmount(feesArray['uluna'], TERRA_DECIMAL)
-    return {
-      type: FeeType.FlatFee,
-      average: baseFeeInLuna,
-      fast: baseFeeInLuna,
-      fastest: baseFeeInLuna,
-    }
+  /**
+   * Get estimated fee.
+   *
+   * @param {Asset} feeAsset Asset to pay fees
+   * @param {CreateTxOptions} options Options to create a simulated tx to estimate fees
+   * @returns {EstimatedFee} Estimated fee
+   */
+  async getEstimatedFee(params: FeeParams): Promise<EstimatedFee> {
+    if (!params) throw new Error('Params need to be passed')
+
+    const { feeAsset, sender, recipient, asset, amount, memo } = params
+
+    const config = this.config[this.network]
+    return await getEstimatedFee({
+      chainId: config.chainID,
+      cosmosAPIURL: config.cosmosAPIURL,
+      sender,
+      recipient,
+      amount,
+      asset,
+      feeAsset,
+      memo,
+      network: this.network,
+    })
   }
+
+  /**
+   * Get fees.
+   *
+   * @param {FeeParams} Fee params (required - they are defined as optional function parameters to fit XChainClient interface only
+   * @returns {Fees} The average/fast/fastest fees.
+   */
+  async getFees(params?: FeeParams): Promise<Fees> {
+    if (!params) throw new Error('Params need to be passed')
+    const { amount: feeAmount } = await this.getEstimatedFee(params)
+    return singleFee(FeeType.PerByte, feeAmount)
+  }
+
   getAddress(walletIndex = 0): string {
     const mnemonicKey = new MnemonicKey({ mnemonic: this.phrase, index: walletIndex })
     return mnemonicKey.accAddress
@@ -129,10 +143,10 @@ class Client extends BaseXChainClient implements XChainClient {
   async getBalance(address: string, assets?: Asset[]): Promise<Balance[]> {
     let balances: Balance[] = []
     let [coins, pagination] = await this.lcdClient.bank.balance(address)
-    balances = balances.concat(this.coinsToBalances(coins))
+    balances = balances.concat(coinsToBalances(coins))
     while (pagination.next_key) {
       ;[coins, pagination] = await this.lcdClient.bank.balance(address, { 'pagination.offset': pagination.next_key })
-      balances = balances.concat(this.coinsToBalances(coins))
+      balances = balances.concat(coinsToBalances(coins))
     }
 
     if (assets) {
@@ -145,13 +159,17 @@ class Client extends BaseXChainClient implements XChainClient {
       return balances
     }
   }
+
   setNetwork(network: Network): void {
     super.setNetwork(network)
     this.lcdClient = new LCDClient({
       URL: this.config[this.network].cosmosAPIURL,
-      chainID: this.config[this.network].ChainID,
+      chainID: this.config[this.network].chainID,
     })
   }
+
+  getConfig = (): ClientConfig => this.config[this.network]
+
   async getTransactions(params?: TxHistoryParams): Promise<TxsPage> {
     //TODO filter by start time?
     //TODO filter by asset
@@ -176,50 +194,93 @@ class Client extends BaseXChainClient implements XChainClient {
     return this.convertTxInfoToTx(txInfo)
   }
 
-  async transfer({ walletIndex = 0, asset = AssetLUNA, amount, recipient, memo }: TxParams): Promise<string> {
+  /**
+   * Transfer
+   *
+   * Note: For paying fees in other than given `asset` (which is used for paying fee by default),
+   * `feeAsset`, `feeAmount`, `gasLimit` parameter are required.
+   * to get all needed data for these three parameters ^,
+   * use `getEstimatedFee` helper from `utils`.
+   *
+   * By passing just `feeAsset` (but not `feeAmount` and `gasLimit`), `feeAmount` and `gasLimit` will be calcaluated internally in `transfer` method.
+   */
+  async transfer({
+    walletIndex = 0,
+    asset = AssetLUNA,
+    amount,
+    recipient,
+    memo,
+    feeAsset,
+    feeAmount,
+    gasLimit,
+  }: TxParams & { feeAmount?: BaseAmount; feeAsset?: Asset; gasLimit?: BigNumber }): Promise<TxHash> {
     if (!this.validateAddress(recipient)) throw new Error(`${recipient} is not a valid terra address`)
 
-    const terraMicroDenom = getTerraMicroDenom(asset.symbol)
-    if (!terraMicroDenom) throw new Error(`${assetToString(asset)} is not a valid terra chain asset`)
+    const assetDenom = getTerraNativeDenom(asset)
+    if (!assetDenom)
+      throw Error(`Invalid asset ${assetToString(asset)} - Only Terra native asset are supported to transfer`)
+
+    const _feeAsset = feeAsset || asset
+    const feeDenom = getTerraNativeDenom(_feeAsset)
+    if (!feeDenom)
+      throw Error(`Invalid asset ${assetToString(_feeAsset)} - Only Terra native assets are supported to pay fees`)
 
     const mnemonicKey = new MnemonicKey({ mnemonic: this.phrase, index: walletIndex })
     const wallet = this.lcdClient.wallet(mnemonicKey)
     const amountToSend: Coins.Input = {
-      [terraMicroDenom]: `${amount.amount().toFixed()}`,
+      [assetDenom]: amount.amount().toFixed(),
     }
-    const send = new MsgSend(wallet.key.accAddress, recipient, amountToSend)
-    const tx = await wallet.createAndSignTx({ msgs: [send], memo })
-    const result = await this.lcdClient.tx.broadcast(tx)
+    const sender = wallet.key.accAddress
+    const send = new MsgSend(sender, recipient, amountToSend)
+
+    let _feeAmount = feeAmount
+    let _gasLimit = gasLimit
+    // Estimate fee amount + gas limit if not passed as params
+    if (!_feeAmount && !_gasLimit) {
+      const { chainID, cosmosAPIURL } = this.getConfig()
+      const { amount: feeAmount, gasLimit } = await getEstimatedFee({
+        chainId: chainID,
+        cosmosAPIURL,
+        sender,
+        recipient,
+        amount,
+        asset,
+        feeAsset: _feeAsset,
+        memo,
+        network: this.getNetwork(),
+      })
+
+      _feeAmount = feeAmount
+      _gasLimit = gasLimit
+    }
+
+    if (!_feeAmount || !_gasLimit) throw Error(`Missing fee amount and/or gas limit`)
+
+    const gasFee: Coin.Data = { amount: _feeAmount.amount().toFixed(), denom: feeDenom }
+    const gasCoins = new Coins([Coin.fromData(gasFee)])
+    const fee = new Fee(_gasLimit.toNumber(), gasCoins)
+
+    const txOptions: CreateTxOptions = {
+      msgs: [send],
+      memo,
+      feeDenoms: [feeDenom],
+      fee,
+    }
+
+    const tx = await wallet.createAndSignTx(txOptions)
+    // broadcast (`async` mode)
+    const result = await this.lcdClient.tx.broadcastAsync(tx)
     return result.txhash
   }
-  private getTerraNativeAsset(denom: string): Asset {
-    if (denom.toLowerCase().includes('luna')) {
-      return AssetLUNA
-    } else {
-      // native coins other than luna, UST, KRT, etc
-      // NOTE: https://docs.terra.money/docs/develop/module-specifications/README.html#currency-denominations
-      const standardDenom = denom.toUpperCase().slice(1, 3) + 'T'
-      return {
-        chain: Chain.Terra,
-        symbol: standardDenom,
-        ticker: standardDenom,
-        synth: false,
-      }
-    }
-  }
 
-  private coinsToBalances(coins: Coins): Balance[] {
-    return (coins.toArray().map((c: Coin) => {
-      return {
-        asset: this.getTerraNativeAsset(c.denom),
-        amount: baseAmount(c.amount.toFixed(), TERRA_DECIMAL),
-      }
-    }) as unknown) as Balance[]
-  }
-
+  // TODO (xchain-contributors) Extract to `util` + add tests
+  // (@xchain-contributors) TODO: Fix `tx` type to avoid `any`
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private convertSearchResultTxToTx(tx: any): Tx {
     let from: TxFrom[] = []
     let to: TxTo[] = []
+    // (@xchain-contributors) TODO: Fix `msg` type to avoid `any`
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     tx.tx.value.msg.forEach((msg: any) => {
       if (msg.type === 'bank/MsgSend') {
         const xfers = this.convertMsgSend(MsgSend.fromAmino(msg))
@@ -247,6 +308,8 @@ class Client extends BaseXChainClient implements XChainClient {
       hash: tx.txhash,
     }
   }
+
+  // TODO (xchain-contributors) Extract to `util` + add tests
   private convertTxInfoToTx(txInfo: TxInfo): Tx {
     let from: TxFrom[] = []
     let to: TxTo[] = []
@@ -278,13 +341,15 @@ class Client extends BaseXChainClient implements XChainClient {
       hash: txInfo.txhash,
     }
   }
+
+  // TODO (xchain-contributors) Extract to `util` + add tests
   private convertMsgSend(msgSend: MsgSend): { from: TxFrom[]; to: TxTo[] } {
     const from: TxFrom[] = []
     const to: TxTo[] = []
     msgSend.amount.toArray().forEach((coin) => {
       //ensure this is in base units ex uluna, uusd
       const baseCoin = coin.toIntCoin()
-      const asset = this.getTerraNativeAsset(baseCoin.denom)
+      const asset = getTerraNativeAsset(baseCoin.denom)
       const amount = baseAmount(baseCoin.amount.toFixed(), TERRA_DECIMAL)
       if (asset) {
         // NOTE: this will only populate native terra Assets
@@ -303,6 +368,8 @@ class Client extends BaseXChainClient implements XChainClient {
 
     return { from, to }
   }
+
+  // TODO (xchain-contributors) Extract to `util` + add tests
   private convertMsgMultiSend(msgMultiSend: MsgMultiSend): { from: TxFrom[]; to: TxTo[] } {
     const from: TxFrom[] = []
     const to: TxTo[] = []
@@ -310,7 +377,7 @@ class Client extends BaseXChainClient implements XChainClient {
       input.coins.toArray().forEach((coin) => {
         //ensure this is in base units ex uluna, uusd
         const baseCoin = coin.toIntCoin()
-        const asset = this.getTerraNativeAsset(baseCoin.denom)
+        const asset = getTerraNativeAsset(baseCoin.denom)
         const amount = baseAmount(baseCoin.amount.toFixed(), TERRA_DECIMAL)
         if (asset) {
           // NOTE: this will only populate native terra Assets
@@ -326,8 +393,9 @@ class Client extends BaseXChainClient implements XChainClient {
       output.coins.toArray().forEach((coin) => {
         //ensure this is in base units ex uluna, uusd
         const baseCoin = coin.toIntCoin()
-        const asset = this.getTerraNativeAsset(baseCoin.denom)
+        const asset = getTerraNativeAsset(baseCoin.denom)
         const amount = baseAmount(baseCoin.amount.toFixed(), TERRA_DECIMAL)
+        // TODO (xchain-contributors) Double check: Why is check of asset needed? `asset` will never fail ...
         if (asset) {
           // NOTE: this will only populate native terra Assets
           to.push({
@@ -340,6 +408,36 @@ class Client extends BaseXChainClient implements XChainClient {
     })
 
     return { from, to }
+  }
+
+  /**
+   * Transaction to THORChain inbound address.
+   *
+   * @param {DepositParams} params The transaction options.
+   * @returns {TxHash} The transaction hash.
+   *
+   * @throws {"halted chain"} Thrown if chain is halted.
+   * @throws {"halted trading"} Thrown if trading is halted.
+   */
+  async deposit({ walletIndex = 0, asset = AssetLUNA, amount, memo }: DepositParams): Promise<TxHash> {
+    const inboundDetails = await getInboundDetails(asset.chain, this.network)
+
+    if (inboundDetails.haltedChain) {
+      throw new Error(`Halted chain for ${assetToString(asset)}`)
+    }
+    if (inboundDetails.haltedTrading) {
+      throw new Error(`Halted trading for ${assetToString(asset)}`)
+    }
+
+    const txHash = await this.transfer({
+      walletIndex,
+      asset,
+      amount,
+      recipient: inboundDetails.vault,
+      memo,
+    })
+
+    return txHash
   }
 }