+++ packages/xchain-util/src/asset.ts
@@ -200,13 +200,6 @@ export const AssetLTC: Asset = { chain: Chain.Litecoin, symbol: 'LTC', ticker: '
  */
 export const AssetDOGE: Asset = { chain: Chain.Doge, symbol: 'DOGE', ticker: 'DOGE', synth: false }
 
-/**
- * Luna asset
- * Based on definition in Thorchain
- * @see TBD
- */
-export const AssetLUNA: Asset = { chain: Chain.Terra, symbol: 'LUNA', ticker: 'LUNA', synth: false }
-
 export const RUNE_TICKER = 'RUNE'
 
 /**
@@ -366,7 +359,7 @@ export const currencySymbolByAsset = ({ ticker }: Asset) => {
       return AssetCurrencySymbol.BTC
     case ticker === AssetETH.ticker:
       return AssetCurrencySymbol.ETH
-    case ticker.includes('USD'):
+    case ticker.includes('USD') || ticker.includes('UST'):
       return AssetCurrencySymbol.USD
     default:
       return ticker
@@ -447,3 +440,12 @@ export const formatBaseAsAssetAmount = ({
   decimal?: number
   trimZeros?: boolean
 }) => formatAssetAmount({ amount: baseToAsset(amount), decimal, trimZeros })
+
+/**
+ * Checks equality of two `Assets`
+ * @param {Asset} a Asset one
+ * @param {Asset} b Asset two
+ * @return {boolean} Result of equality check
+ */
+export const eqAsset = (a: Asset, b: Asset) =>
+  a.chain === b.chain && a.symbol === b.symbol && a.ticker === b.ticker && a.synth === b.synth