+++ packages/xchain-doge/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v.0.2.1 (2022-05-05)
+
+## Update
+
+- Add `deposit` function to Doge `Client`
+- Update latest dependencies
+- Add tests for `deposit` 
+
+## Fix
+
+- Fix import of `xchain-client`
+
 # v.0.2.0 (2022-03-23)
 
 ## Update